Renaming generated graphql schema file for the shared/components library to correlate to other generated schema files and to be ignored from git.
